typescript code fixes
