+++ components/about-section-bucket.tsx
@@ -1,6 +1,35 @@
 import React from 'react';
 import parse from 'html-react-parser';
-import { BucketProps, Bucket } from "../typescript/about-section-bucket";
+import { Action,Image } from '../typescript/action';
+
+type AdditionalParam = {
+  title_h2?: string;
+  title_h3?: string;
+  description?: string;
+}
+
+type Bucket = {
+  title_h3: string;
+  description: string;
+  icon: Image;
+  $: AdditionalParam;
+  url: string;
+}
+
+type BucketsList = {
+  title_h3: string;
+  description: string;
+  url: string;
+  call_to_action: Action;
+  icon: Image;
+  $: AdditionalParam;
+}
+
+type BucketProps = {
+  title_h2: string;
+  buckets:[BucketsList];
+  $: AdditionalParam;
+}
 
 export default function AboutSectionBucket({ sectionWithBuckets }: {sectionWithBuckets:BucketProps}) {
   function bucketContent(bucket: Bucket, index: number) {
