+++ components/blog-banner.tsx
@@ -1,5 +1,20 @@
 import React from 'react';
-import {BannerProps} from '../typescript/blog';
+
+type AdditionalParam = {
+  banner_title:string;
+  banner_description: string;
+  title: {};
+  title_h2: string;
+  body: string;
+  date: string;
+}
+
+type BannerProps = {
+  banner_title:string;
+  banner_description: string;
+  bg_color: string;
+  $: AdditionalParam;
+}
 
 export default function BlogBanner({ blogBanner }: {blogBanner : BannerProps}) {
   return (
