+++ components/blog-list.tsx
@@ -2,12 +2,37 @@ import React from 'react';
 import moment from 'moment';
 import parse from 'html-react-parser';
 import Link from 'next/link';
-import { BloglistProps } from "../typescript/blog";
+import { Image } from "../typescript/action";
+
+type AdditionalParam = {
+  banner_title:string;
+  banner_description: string;
+  title: {};
+  title_h2: string;
+  body: string;
+  date: string;
+}
+
+type Author = {
+  title: string;
+  $: AdditionalParam;
+}
+
+
+type BloglistProps = {
+  body: string;
+  url: string;
+  featured_image: Image; 
+  title: string;
+  date: string;
+  author: [Author];
+  $: AdditionalParam;
+}
 
 function BlogList({ bloglist }: { bloglist: BloglistProps }) {
-  let body = typeof bloglist.body === 'string' && bloglist.body.substr(0, 300);
-  const stringLength = (body as string).lastIndexOf(' ');
-  body = `${(body as string).substr(0, Math.min((body as string).length, stringLength))}...`;
+  let body: string = bloglist.body && bloglist.body.substr(0, 300);
+  const stringLength = body.lastIndexOf(' ');
+  body = `${body.substr(0, Math.min(body.length, stringLength))}...`;
   return (
     <div className='blog-list'>
       {bloglist.featured_image && (
