+++ components/layout.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Header from './header';
 import Footer from './footer';
 import DevTools from './devtools';
-import { HeaderProps, FooterProps, PageProps, Posts, ChilderenProps, Entry, Links } from "../typescript/layout";
+import { HeaderProps, FooterProps, PageProps, Posts, ChilderenProps, Entry, NavLinks, Links } from "../typescript/layout";
 
 export default function Layout({
   header,
@@ -20,13 +20,13 @@ export default function Layout({
   blogPost && (jsonObj.blog_post = blogPost);
   blogList && (jsonObj.blog_post = blogList);
 
-  function buildNavigation(ent: Entry, hd: any, ft: any) {
+  function buildNavigation(ent: Entry, hd: HeaderProps, ft: FooterProps) {
     let newHeader = { ...hd };
     let newFooter = { ...ft };
     if (ent.length !== newHeader.navigation_menu.length) {
       ent.forEach((entry) => {
         const hFound = newHeader?.navigation_menu.find(
-          (navLink: Links) => navLink.label === entry.title
+          (navLink: NavLinks) => navLink.label === entry.title
         );
         if (!hFound) {
           newHeader.navigation_menu?.push({
