+++ components/section.tsx
@@ -1,6 +1,25 @@
 import React from 'react';
 import Link from 'next/link';
-import { SectionProps } from "../typescript/section";
+import { Image, Action } from "../typescript/action";
+
+type AdditionalParam = {
+  title: string;
+  title_h2: string;
+  title_h3: string;
+  description: string;
+  html_code: string;
+  designation: string;
+  name: string;
+}
+
+type SectionProps = {
+  title_h2: String;
+  description: string;
+  call_to_action: Action;
+  image: Image;
+  image_alignment: string;
+  $: AdditionalParam;
+} 
 
 export default function Section({ section }: {section : SectionProps}) {
   function contentSection(key: any) {
