+++ components/team-section.tsx
@@ -1,5 +1,29 @@
 import React from 'react';
-import { TeamProps } from "../typescript/section";
+import { Image } from "../typescript/action";
+
+type AdditionalParam = {
+  title: string;
+  title_h2: string;
+  title_h3: string;
+  description: string;
+  html_code: string;
+  designation: string;
+  name: string;
+}
+
+type Employee = {
+  image: Image;
+  name: string;
+  designation: string;
+  $: AdditionalParam;
+}
+
+type TeamProps = {
+  title_h2: string;
+  description: string;
+  $: AdditionalParam;
+  employees: [Employee];
+}
 
 export default function TeamSection({ ourTeam }: {ourTeam : TeamProps}) {
   return (
