--- typescript/archive-relative.ts
@@ -1,15 +1,0 @@
-type Object = {
-    title: string;
-    body: string;
-  }
-  
-  type Blog = {
-    url: string;
-    body: string;
-    title: string;
-    $: Object;
-  }
-  
-export type BlogListProps = {
-    blogs: [Blog];
-  }
