+++ typescript/layout.ts
@@ -1,7 +1,7 @@
 import { Image } from "./action";
 import { Component } from "../typescript/component";
 
-type Object = {
+type AdditionalParam = {
   title: {};
   copyright: string;
   announcement_text: string;
@@ -12,29 +12,24 @@ type Object = {
 type EntryData = {
   title: string;
   url: string;
-  $: Object;
+  $: AdditionalParam;
 }
 
 type Announcement = {
   show_announcement: boolean;
   announcement_text: string;
-  $: Object;
+  $: AdditionalParam;
 }
 
 type PageRef = {
+  title: string;
   url: string;
-  $: Object;
-}
-
-type List = {
-  label: string;
-  page_reference: [PageRef];
-  $: Object;
+  $: AdditionalParam;
 }
 
 type Share = {
   link: Links;
-  icon: Image
+  icon: Image;
 }
 
 type Social = {
@@ -47,14 +42,14 @@ type Navigation = {
 
 type Author = {
   title: string;
-  $: Object;
+  $: AdditionalParam;
 }
 
 type Blog = {
   url: string;
   body: string;
   title: string;
-  $: Object;
+  $: AdditionalParam;
 }
 
 export type Posts = {
@@ -71,24 +66,40 @@ export type Posts = {
   _owner: {}
 }
 
+
 export type HeaderProps = {
   locale:string;
   logo: Image;
   navigation_menu:[List]
-  notification_bar: Announcement
+  notification_bar: Announcement;
   title: string;
   uid: string;
+  social: Social;
+  navigation: Navigation;
+  copyright: string;
+  $: AdditionalParam;
 }
 
 export type Entry = [
   entry: EntryData
 ]
 
+type List = {
+  label?: string;
+  page_reference: [PageRef];
+  $: {};
+  href?: string;
+}
+
+export type NavLinks = {
+  label?: string;
+}
+
 export type Links = {
-  label: string;
+  label?: string;
   title: string;
   href: string;
-  $:Object;
+  $:AdditionalParam;
 }
 
 export type PageProps = {
@@ -106,7 +117,11 @@ export type FooterProps = {
   social: Social;
   navigation: Navigation;
   copyright: string;
-  $: Object;
+  locale: string, 
+  navigation_menu: [List];
+  notification_bar: Announcement; 
+  uid: string;
+  $: AdditionalParam;
 }
 
 export type ChilderenProps = {
